Grammar in read.Rmd
`store them file in variables` -> `store them in variables`

I've also suggested replacing `in Terminal` with `in a terminal`, since the former reads as if the user is in a `*nix` environment with the "Terminal" application (e.g. in MacOS or Linux), but users may be following along in Windows with the command prompt, PowerShell, etc.